Move Tooltip code to future internally
