+++ .changeset/strange-readers-talk.md
@@ -1,0 +1,5 @@
+---
+'@kaizen/components': patch
+---
+
+Move Tooltip code to future internally
