+++ packages/components/src/Focusable/_docs/ApiSpecification.mdx
@@ -2,10 +2,10 @@ import { Canvas, Controls, Meta } from '@storybook/blocks'
 import { KAIOInstallation, ResourceLinks, SbContent } from '~storybook/components'
 import * as exampleStories from './Focusable.stories'
 
-<Meta title="Overlays/Focusable/v3/API Specification" />
+<Meta title="Components/Focusable/API Specification" />
 
 <SbContent>
-  # Focusable API Specification (v3)
+  # Focusable API Specification
 
 Updated June 25, 2024
 
@@ -17,7 +17,7 @@ Updated June 25, 2024
 />
 
 <SbContent className="mb-24">
-  <KAIOInstallation exportNames={["Focusable"]} family="overlays" version="3" />
+  <KAIOInstallation exportNames={["Focusable"]} />
 
 ## Overview
