+++ packages/components/src/Tooltip/_docs/Tooltip.mdx
@@ -4,7 +4,7 @@ import * as TooltipStories from './Tooltip.stories'
 
 <Meta of={TooltipStories} />
 
-# Tooltip (v1)
+# Tooltip
 
 <ResourceLinks
   sourceCode="https://github.com/cultureamp/kaizen-design-system/tree/main/packages/components/src/Tooltip"
@@ -15,10 +15,8 @@ import * as TooltipStories from './Tooltip.stories'
 
 <Installation
   installCommand="pnpm add @kaizen/components"
-  importStatement='import { Tooltip } from "@kaizen/components/v1/overlays"'
+  importStatement='import { Tooltip } from "@kaizen/components"'
 />
-
-It is recommended that you import from the `v2` entry point.
 
 ## Overview
