+++ packages/components/src/Tooltip/_variables.scss
