+++ packages/components/src/__future__/Tooltip/_docs/ApiSpecification.mdx
@@ -3,10 +3,10 @@ import { KAIOInstallation, ResourceLinks, SbContent } from '~storybook/component
 import * as docsStories from './Tooltip.docs.stories'
 import * as exampleStories from './Tooltip.stories'
 
-<Meta title="Overlays/Tooltip/v3/API Specification" />
+<Meta title="Components/Tooltip/Tooltip (Future)/API Specification" />
 
 <SbContent>
-  # Tooltip API Specification (v3)
+  # Tooltip API Specification (Future)
 
 Updated June 25, 2024
 
@@ -23,7 +23,7 @@ Updated June 25, 2024
 
 {' '}
 
-<KAIOInstallation exportNames={['Tooltip', 'TooltipTrigger']} family="overlays" version="3" />
+<KAIOInstallation exportNames={['Tooltip', 'TooltipTrigger']} />
 
 ## Overview
