+++ packages/components/src/__future__/Tooltip/_docs/Tooltip.mdx
@@ -12,10 +12,10 @@ import PlacementPng from './assets/tooltip_placement.png'
 import DesignSpecPng from './assets/tooltip_spec.png'
 import WhenToUsePng from './assets/tooltip_variant.png'
 
-<Meta title="Overlays/Tooltip/v3/Usage Guidelines" />
+<Meta title="Components/Tooltip/Tooltip (Future)/Usage Guidelines" />
 
 <SbContent>
-  # Tooltip (v3)
+  # Tooltip (Future)
 
 Updated June 25, 2024
