+++ packages/components/src/__future__/Tooltip/index.ts
@@ -1,0 +1,2 @@
+export * from './Tooltip'
+export * from './TooltipTrigger'
