+++ packages/components/src/__future__/index.ts
@@ -2,3 +2,4 @@ export * from './Icon'
 export * from './Select'
 export * from './Tag'
 export * from './Tabs'
+export * from './Tooltip'
