+++ packages/components/src/__overlays__/Focusable/index.ts
@@ -1,0 +1,1 @@
+export * from '../../Focusable'
