+++ packages/components/src/__overlays__/Tooltip/v2/index.ts
@@ -1,1 +1,1 @@
-export * from '../v1/Tooltip'
+export * from '../../../Tooltip'
