+++ packages/components/src/__overlays__/Tooltip/v3/index.ts
@@ -1,2 +1,1 @@
-export * from './Tooltip'
-export * from './TooltipTrigger'
+export * from '../../../__future__/Tooltip'
