+++ packages/components/src/index.ts
@@ -20,6 +20,7 @@ export * from './ErrorPage'
 export * from './FieldGroup'
 export * from './FieldMessage'
 export * from './Filter'
+export * from './Focusable'
 export * from './GuidanceBlock'
 export * from './Heading'
 export * from './Icon'
@@ -58,6 +59,6 @@ export * from './utils'
 export * from './VisuallyHidden'
 export * from './Well'
 export * from './Workflow'
-export * from './__overlays__/Tooltip/v1'
+export * from './Tooltip'
 export * from './__actions__/Menu/v1'
 export * from './__actions__/Button/v1'
